### 2018-07-24 3fc4f7b
add missing figure scaling.pdf
