### 2018-07-24 #831 @pdebuyl
From the list of reviewers, I can suggest:
- @tacaswell
- @richardjgowers
- @ha0ye
- @malramsay64

If this is not asking too much, can we ask whedon to pick another citation style? See issue https://github.com/openjournals/whedon/issues/2
