### 2018-07-30 #831 @richardjgowers
I’m happy to review this, or is it for the editor to pick?
