### 2018-07-30 #831 @pdebuyl
@arfon thank you for the update. No worry with the delay, I understand editorial loads and the increase for JOSS just means more work from volunteers.
