### 2018-08-06 #877 @whedon
[REVIEW]: 'tidynamics: A tiny package to compute the dynamics of stochastic and molecular simulations
**Submitting author:** @pdebuyl (<a href="http://orcid.org/0000-0002-6640-6463">Pierre de Buyl</a>)
**Repository:** <a href="https://github.com/pdebuyl-lab/tidynamics" target ="_blank">https://github.com/pdebuyl-lab/tidynamics</a>
**Version:** 1.0.0
