### 2018-08-06 #877 @kyleniemeyer
@pdebuyl @richardjgowers @malramsay64 👋
