### 2018-08-06 52956b7
Surround version import in try/except block
This defaults the __version__ to 'dev' should the version.py file not be
found on import. This occurs when importing tidynamics from the
source directory. This commit fixes #3.
