+++ tidynamics/__init__.py
@@ -1,3 +1,7 @@
 
 from ._correlation import acf, msd, cross_displacement, correlation
-from .version import __version__ as __version__
+
+try:
+    from .version import __version__ as __version__
+except ModuleNotFoundError as e:
+    __version__ = "dev"
