### 2018-08-06 #877 @malramsay64
@kyleniemeyer I can see that @pdebuyl is preparing for release 0.1.2, which I assume will be in concert with the completion of the review, yet the version is not technically 0.1.2 being a pre-release version (0.1.2.dev5). Is a 0.1.2 release required for review? Also does there have to be a GitHub release (tagged commit) for the version being reviewed?
