### 2018-08-06 #877 @kyleniemeyer
@malramsay64 we don't have a requirement on this—I think it's fine to hold the 0.1.2 release pending for any changes that come about due to the JOSS review. I will expect that version to be released and archived when the submission in accepted, though
