### 2018-08-08 pdebuyl-lab/tidynamics#8 @malramsay64
[review] Improve API Documentation
I would consider the certain parts of the documentation to be inadequate for giving me a good grasp of using the API. As a reference for what I would expect, the documentation for the distance functions in [MDAnalysis](https://www.mdanalysis.org/docs/documentation_pages/analysis/distances.html) where there are few sentences providing helpful information on using the function. For explaining how to use the software I think your [RMPCDMD](http://lab.pdebuyl.be/rmpcdmd/algorithms.html) documentation does a good job.

Some specific points which I think need addressing:

- [x] The only information about the `tidynamics.cross_correlation` function is that it performs a cross correlation. Correlation and MSD is mentioned in the [Algorithms section](http://lab.pdebuyl.be/tidynamics/algorithms.html)
- [x] I find the documentation of the [data format](http://lab.pdebuyl.be/tidynamics/data_format.html) confusing. An example would go a long way in helping the explanation.
- [x] Have examples which are separate from the generation of the figures, since it is unclear what parts are for the figures and what is required to use the package. [This line](https://github.com/pdebuyl-lab/tidynamics/blob/3fc4f7b1b96f47eeb69f6fd0e9c57f25e795ea79/examples/plot_msd_1.py#L45) is a good example, do I need to discard half the results, or is that just for the figure.
