### 2018-08-08 pdebuyl-lab/tidynamics#9 @malramsay64
[review] Include matplotlib and pytest dependencies in instructions
The instructions for running the tests assume that the pytest pacakge is installed. Additionally the examples assume matplotlib is installed. I would consider a note with install instructions that these are required to be adequate.
