### 2018-08-08 pdebuyl-lab/tidynamics#10 @malramsay64
[review] Examples with no additional dependencies
Currently running any of the examples requires the additional installation of matplotlib. There should be at least *an* example that can be run without installing additional dependencies.
