### 2018-08-08 #877 @malramsay64
@pdebuyl I have mostly finished my review, with mostly the assessment of functionality remaining. I have raised a number of issues in the repository, with issues marked [review] specifically not meeting review criteria. These are
- pdebuyl-lab/tidynamics#8
- pdebuyl-lab/tidynamics#9
- pdebuyl-lab/tidynamics#10

The remaining issues I have raised are more suggestions.

Some general notes.
- I like the concept of the library and would like to use it for my research, however I don't find the documentation particularly helpful for understanding how to use the software.
- There appears to be a lack of consistency throughout the package.
- With a stated goal of being a reference implementation I would expect more comprehensive documentation.

I'm happy to discuss/help with anything I have raised.
