### 2018-08-11 #877 @richardjgowers
@pdebuyl : @malramsay64 has made good points about docs.  I think the tests should also be a bit more comprehensive, ideally some realistic data but at least some seeded random walks to act as regression tests.
