### 2018-08-12 daf9c4b
use consistently single quotes
an instance of double quotes remains to escape the single quotes inside

fixes #5
