### 2018-08-12 5ee277e
replace np.array by np.asarray in _correlation.py
fixes #6
