### 2018-08-12 2da22ae
change argument of cross_correlation from x to pos
fixes #11
