### 2018-08-12 1371102
add link to CONTRIBUTORS in README.rst
