### 2018-08-12 bc1a3b7
mention conda-forge as installation method
fixes #7
