+++ README.rst
@@ -35,6 +35,10 @@ tidynamics can be installed with pip::
 
     pip install --user tidynamics
 
+or with conda (via conda-forge)::
+
+    conda install -c conda-forge tidynamics
+
 It is also possible to download the source code and execute the setup.py file.
 
 I ran the tests with Python 2.7, 3.5 and 3.6 and NumPy 1.11 and 1.13. If you encounter any
