### 2018-08-12 pdebuyl-lab/tidynamics#9 @pdebuyl
I think of adding a note on installing pytest next to

> Testing
> We use pytest for testing:
>
> python -m pytest

in the README and doc, and one on installing matplotlib in the text introducing the examples here: http://lab.pdebuyl.be/tidynamics/auto_examples/index.html This text is stored in `examples/README.txt` so it would be visible from the documentation as well as from the repository.
