### 2018-08-12 pdebuyl-lab/tidynamics#9 @malramsay64
Sounds good to me.
