### 2018-08-13 533d045
add instructions to run the tests
download the source
install pytest

fixed #9
