### 2018-08-13 #877 @pdebuyl
Thanks all for the feedback. I have started to work on the issues.
