### 2018-08-13 e9c0cc4
add standalone cli tool and example data
fixes #10
