### 2018-08-13 pdebuyl-lab/tidynamics#10 @pdebuyl
I added a standalone command-line tool `examples/tidynamics_tool.py` with corresponding sample data to address the issue. The tool will likely be useful to CLI enthusiasts :-)
