### 2018-08-13 pdebuyl-lab/tidynamics#10 @pdebuyl
I updated the corresponding documentation page http://lab.pdebuyl.be/tidynamics/auto_examples/index.html
