### 2018-08-13 a08061a
fix inconsistent docstring of msd
