+++ doc/bibtex.bib
@@ -12,3 +12,16 @@
 	abstract = "nMOLDYN is a modular program package for the analysis of Molecular Dynamics simulations. It is especially designed for the computation and decomposition of neutron scattering spectra. The structure and dynamics of the simulated systems can be characterized in terms of various space and time correlation functions. To analyze the dynamics of complex systems, rigid-body motions of arbitrarily chosen molecular subunits can be studied. All modules use a common input data structure for the MD trajectory files which facilitates an efficient calculation of time correlation functions by Fast Fourier Transform algorithms."
 }
 
+
+@article{kraft_complex_shape_2013,
+	keywords = {colloid, hydrodynamics, brownian},
+	file = {dossiers/ebib/kraft_wittkowski_ten_hagen_edmond_pine_lowen_brownian_motion_and_the_hydrodynamic_friction_tensor_for_colloidal_particles_of_complex_shape_pre_2013.pdf},
+	title = {Brownian motion and the hydrodynamic friction tensor for colloidal particles of complex shape},
+	author = {Kraft, Daniela J. and Wittkowski, Raphael and ten Hagen, Borge and Edmond, Kazem V. and Pine, David J. and L\"owen, Hartmut},
+	journal = {Phys. Rev. E},
+	volume = {88},
+	pages = {050301},
+	year = {2013},
+	doi = {10.1103/PhysRevE.88.050301}
+}
+
