### 2018-08-14 pdebuyl-lab/tidynamics#8 @pdebuyl
I added a part in "algorithms" to describe the `cross_displacement` function.

I edited both the documentation and the docstrings to improve the consistency. I don't know what to add to the dosctrings however, as the operations "autocorrelation" and "mean-square displacement" are standard and I do not provided options that change their behaviors, as it is the case in MDAnalysis' documentation.

I believe this addresses points 1 and 3. I am waiting for your feedback before doing 2.
