+++ setup.py
@@ -18,7 +18,7 @@ setup(name='tidynamics',
       author='Pierre de Buyl',
       author_email='pdebuyl@pdebuyl.be',
       license='BSD',
-      url='https://pypi.python.org/pypi/tidynamics',
+      url='https://pypi.org/project/tidynamics/',
       packages=['tidynamics'],
       install_requires=['numpy'],
       classifiers=[
