### 2018-08-14 2901b69
centralize version in file VERSION
