+++ doc/conf.py
@@ -20,7 +20,8 @@
 # import os
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
-
+import os.path
+import re
 
 # -- General configuration ------------------------------------------------
 
@@ -67,10 +68,13 @@ author = 'Pierre de Buyl'
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
+with open(os.path.join(os.path.pardir, 'tidynamics', 'VERSION')) as f:
+    # The full version, including alpha/beta/rc tags.
+    release = f.read().strip()
+
 # The short X.Y version.
-version = '0.1.2'
-# The full version, including alpha/beta/rc tags.
-release = '0.1.2.dev5'
+version = re.match('\d+\.\d+\.\d+', release).group()
+
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
