### 2018-08-15 pdebuyl-lab/tidynamics#8 @malramsay64
Those changes look good.

For the docstrings, would adding something like the below to the mean squared displacement be appropriate?

    This function computes the MSD at each input time for the particle in the input
    trajectory. Note that over short timescales this will give a slightly different
    result to the standard method.

I haven't come across this particular algorithm before so I am not entirely sure what to put there, so feel free to change it. The idea being, what is useful for the person using the function? I completely understand that documentation is hard.

Now you have the two distinct input data types, `data` and `pos` the data format will be more easily differentiated. I think the important point about 2 is the first dimension being time, which I think is why I was initially confused since I am more used to it being a particle index.
