### 2018-08-16 pdebuyl-lab/tidynamics#8 @pdebuyl
The method actually for the correlations and for the mean-square displacement gives identical result to the full-blown O(N²) computation. I will review once more the docstrings with the aim of "obviousness" :-)

Concerning 2, I will add an example to be completely specific.
