### 2018-08-16 pdebuyl-lab/tidynamics#8 @pdebuyl
Update of [Data format](http://lab.pdebuyl.be/tidynamics/data_format.html)
