### 2018-08-16 0969c71
add brute force test for msd
