### 2018-08-16 pdebuyl-lab/tidynamics#8 @malramsay64
I get it now :) The test comparing against to full computation is a great addition 👍 I was just comparing it to each point individually, hence the perceived difference.

As the documentation currently stands I think it is sufficient from the perspective of the review. It is definitely significantly easier to understand.
