### 2018-08-16 pdebuyl-lab/tidynamics#8 @pdebuyl
Thank you for the feedback!
