### 2018-08-17 #877 @pdebuyl
Hi all,

While waiting for the end of the review, and after improving the documentation, I thought of providing an example notebook that one can launch on mybinder.

See https://github.com/pdebuyl-lab/tidynamics/tree/binder_notebook

It is in a separate branch for now, but as it is a coming addition to the software I wanted to notify the review page.

Also, for a first release "0.1.2" is not a nice number and I plan to release 0.2.0 when the review is over and I have to archive the code to zenodo.

Regards,

Pierre
