### 2018-08-20 eb40716
use local version for building the docs
