+++ doc/conf.py
@@ -17,10 +17,10 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
 import os.path
+import sys
+sys.path.insert(0, os.path.pardir)
 import re
 
 # -- General configuration ------------------------------------------------
