### 2018-08-20 0a45308
edit binder link and add it to the doc
