### 2018-08-20 #877 @richardjgowers
@kyleniemeyer ok I'm happy to accept this.  @pdebuyl if you're happy that the API won't change you can just call it 1.0.0 according to [semantic versioning](https://semver.org)
