### 2018-08-20 1a061d6
add release info in index
