### 2018-08-20 #877 @pdebuyl
@kyleniemeyer I believe I addressed all reviewers' feedback.
