### 2018-08-21 #877 @malramsay64
I am also happy to accept this. @pdebuyl I also agree with @richardjgowers on version 1.0.0.
