### 2018-08-21 #877 @pdebuyl
Ok, I'll release as 1.0.0 on zenodo.

@arfon may I include the DOI badge and the corresponding citation already in 1.0.0 (so that the released version is "self-contained")? I plan to add a `CITATION` file and the corresponding information in the README, `doc/index.rst` and `tidynamics/__init__.py` file.

From my understanding the doi is reserved already and visible in the pdf builds from whedon, is that the good one?
