### 2018-08-21 #877 @arfon
> From my understanding the doi is reserved already and visible in the pdf builds from whedon, is that the good one?

@pdebuyl yet that's fine.  The DOI for the paper will be https://doi.org/10.21105/joss.00877
