### 2018-08-21 451e8a2
include DOI in the documentation, add CITATION file
