### 2018-08-21 #877 @pdebuyl
tidynamics 1.0.0 is archived on zenodo as

https://doi.org/10.5281/zenodo.1401185
