### 2018-08-21 #877 @kyleniemeyer
Hi @pdebuyl @arfon, I just need to copy-edit the paper first, then we will be ready to accept.
