### 2018-08-26 #877 @kyleniemeyer
A few comments on the paper:

- The second sentence is a little confusing due to the comma use: "The autocorrelation, the correlation between two time-series, the mean-square displacement, and the cross-displacement (for off-diagonal realisations of Brownian motion), allow...". Perhaps something like this would be more clear? "The autocorrelation (the correlation between two time-series), the mean-square displacement, and the cross-displacement (for off-diagonal realisations of Brownian motion) allow the quantitative measure of relaxation and transport coefficients from numerical trajectories."
- Third paragraph: "Python 2.7, 3.5 and 3.6" -> "Python 2.7, 3.5, and 3.6"
- Fourth paragraph: is the sentence "We show the actual CPU time and compare it favourably to the expected scaling in the example Scaling behaviour." meant to refer to Figure 3?
- Sixth paragraph: "The Fast Correlation Algorithm as applied to mean-square displacements was proposed in (G. R. Kneller et al. 1995)" -> "The Fast Correlation Algorithm as applied to mean-square displacements was proposed by Kneller et al. (1995)". You should be able to do this by replacing `[@nmoldyn_1995]` with just `@nmoldyn_1995`.
- Seventh paragraph: "The benefits of using tidynamics originate in the implementation of the suitable definitions for the study of dynamical systems, in a good performance, and in its ease of installation." -> "The benefits of using tidynamics originate in the implementation of the suitable definitions for the study of dynamical systems, good performance, and its ease of installation."
- In "Examples", please capitalize "figures". Also, the legends in Figures 1 and 3 overlap with the content of the plots. Could you regenerate with them a bit bigger to avoid this?
- Throughout, some of the inline references have authors initials in them—I'm not sure how this happened, but it should just be their last name. Also, https://doi.org/10.1109/MCSE.2007.58 might be a better reference for NumPy

Please let me know when you are finished making these changes.
