### 2018-08-27 33e6aea
take into account feedback from the editor
1. clarity of first paragraph
2. typos and style issues
3. update of the figures
