+++ paper/paper.bib
@@ -62,3 +62,14 @@
 	abstract = {Abstract We present a new version of the program package nMoldyn, which has been originally developed for a neutron-scattering oriented analysis of molecular dynamics simulations of macromolecular systems (Kneller et al., Comput. Phys. Commun. 1995, 91, 191) and was later rewritten to include in-depth time series analyses and a graphical user interface (Rog et al., J. Comput. Chem. 2003, 24, 657). The main improvement in this new version and the focus of this article are the parallelization of all the analysis algorithms for use on multicore desktop computers as well as distributed-memory computing clusters. The parallelization is based on a task farming approach which maintains a simple program structure permitting easy modification and extension of the code to integrate new analysis methods. © 2012 Wiley Periodicals, Inc.}
 }
 
+
+@Article{oliphant_cise_2007,
+	Title = {Python for Scientific Computing},
+	Author = {Travis E Oliphant},
+	Journal = {Computing in Science {\&} Engineering},
+	Year = {2007},
+	Pages = {10--20},
+	Volume = {9},
+	doi = {10.1109/MCSE.2007.58},
+}
+
