### 2018-08-27 #877 @kyleniemeyer
@pdebuyl looks good, thanks! Please now archive this latest version with the updated paper (1.0.1, maybe?) and provide the DOI, and we'll be done.
