### 2018-08-27 #877 @pdebuyl
Hi @kyleniemeyer , thanks for the editing. The doi for 1.0.1 is https://doi.org/10.5281/zenodo.1404296
