### 2018-08-27 #877 @pdebuyl
Thanks @kyleniemeyer @malramsay64 @richardjgowers @arfon for editing and reviewing!
